Use finite state automata to make splitting O(n^2) -> O(n)
The word piece splitting functionality was quadratic time. This change
makes the splitting linear time by using finite state automata.

The prefix and suffixes set are stored in dictionary automata. A new
type, PrefixAutomaton is added. A prefix automaton matches every
prefix in a string. E.g. for the string "Rust", it matches:

[empty prefix]
R
Ru
Rus
Rust

By computing the intersection language of a dictionary automaton and a
prefix automaton, we can find the matching prefixes in the automaton
in linear time. We then take the longest matching prefix as a word
piece.

While at it, also implement TryFrom to construct the finite state
dictionaries from a file.