Find prefixes using binary search
Use a sorted set of strings to represent the word pieces and use
binary search to find the word pieces. This increases the upper bound
from O(N) to O(N log K). However, it is faster in practice due to
avoiding linear traversals of outgoing transitions of a state.

In a quick test on TüBa-D/Z, it's ~5 times faster, while still
avoiding the O(N^2) bound of the most straightforward implementation.